requirements.txt
appdirs==1.4.3
Django==2.0
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
mimeparse==0.1.3
olefile==0.44
packaging==16.8
Pillow==4.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.3
reportlab==3.4.0
six==1.11.0
sqlparse==0.2.4
